+++ vorago-shared-hal/src/uart/tx_async.rs
@@ -214,7 +214,8 @@ impl TxAsync {
     ///
     /// The user MUST ensure that the `Drop` method of all futures generated with this driver
     /// is called on transfer cancellation. By default, this does not require any special handling.
-    pub unsafe fn new(tx: Tx) -> Self {
+    /// This case was considered exotic enough to not justify an `unsafe` API.
+    pub fn new(tx: Tx) -> Self {
         Self(tx)
     }